final fixes

+++ checkers/abstract_checker.h
@@ -8,7 +8,7 @@
 class AbstractChecker {
 public:
 
-    virtual int checkAll(vector<int>) = 0;
+    virtual double checkAll(vector<int>) = 0;
 
     virtual int checkResourceConstraint() = 0;
